# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Poll service.",
        "title": "Poll Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/sheet/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new sheet (verified admin or super admin).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sheets"
                ],
                "summary": "Create sheet",
                "parameters": [
                    {
                        "description": "Sheet creation payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SheetCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.SheetCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sheet/delete": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a sheet by identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sheets"
                ],
                "summary": "Delete sheet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sheet identifier",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sheet/fetch": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve sheets depending on admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sheets"
                ],
                "summary": "List sheets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SheetListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.PaginationResult": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "domain.PollAdminResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "participant": {
                    "type": "integer"
                },
                "poll_type": {
                    "$ref": "#/definitions/domain.pollType"
                },
                "title": {
                    "type": "string"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "domain.Sheet": {
            "type": "object",
            "required": [
                "title",
                "venue"
            ],
            "properties": {
                "approved_at": {
                    "type": "string"
                },
                "approved_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_phone_required": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/domain.SheetStatus"
                },
                "title": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
        "domain.SheetCreatePoll": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "poll_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.SheetCreateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_phone_required": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "polls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SheetCreatePoll"
                    }
                },
                "title": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
        "domain.SheetCreateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "polls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PollAdminResponse"
                    }
                },
                "sheet": {
                    "$ref": "#/definitions/domain.Sheet"
                }
            }
        },
        "domain.SheetListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Sheet"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/domain.PaginationResult"
                }
            }
        },
        "domain.SheetStatus": {
            "type": "string",
            "enum": [
                "pending",
                "published",
                "rejected"
            ],
            "x-enum-varnames": [
                "SheetStatusPending",
                "SheetStatusPublished",
                "SheetStatusRejected"
            ]
        },
...
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.pollType": {
            "type": "string",
            "enum": [
                "single_choice",
                "multi_choice",
                "slide"
            ],
            "x-enum-varnames": [
                "singleChoice",
                "multiChoice",
                "slide"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}